now using @ArquillianSuite for faster tests + EntityIT removed from lib + organized imports
