--- dauphine-open-data-lib/src/test/java/io/github/oliviercailloux/y2018/opendata/ArquillianUtils.java
@@ -1,35 +1,0 @@
-package io.github.oliviercailloux.y2018.opendata;
-
-import java.io.File;
-
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.jboss.shrinkwrap.resolver.api.maven.Maven;
-
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-public class ArquillianUtils {
-
-	public static String makeWarName(final String warName) {
-		return warName + ".war";
-	}
-
-	public static WebArchive makeWar(final String warName) {
-		final File[] libs = Maven.resolver().loadPomFromFile("pom.xml").importRuntimeAndTestDependencies().resolve()
-				.withTransitivity().asFile();
-
-		final JavaArchive jar = ShrinkWrap.create(JavaArchive.class)
-				.addPackages(true, "io.github.oliviercailloux.y2018.opendata.entity")
-				.addAsResource("META-INF/persistence.xml")
-				.addAsResource("log4j2.xml")
-				.addAsResource("arquillian.xml");
-
-		final WebArchive war = ShrinkWrap.create(WebArchive.class, makeWarName(warName)).addAsLibraries(libs)
-				.addAsLibrary(jar);
-
-		return war;
-	}
-
-}
