+++ dauphine-open-data-lib/src/test/java/io/github/oliviercailloux/y2018/opendata/mapper/MapperVcardTest.java
@@ -6,6 +6,7 @@ import org.junit.Test;
 
 import ezvcard.VCard;
 import io.github.oliviercailloux.y2018.opendata.entity.Person;
+import io.github.oliviercailloux.y2018.opendata.mapper.MapperVcard;
 
 /**
  * @author Elhadj Diallo
