+++ dauphine-open-data/src/main/java/io/github/oliviercailloux/y2018/opendata/provider/CORSFilter.java
@@ -1,10 +1,11 @@
 package io.github.oliviercailloux.y2018.opendata.provider;
 
+import java.io.IOException;
+
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerResponseContext;
 import javax.ws.rs.container.ContainerResponseFilter;
 import javax.ws.rs.ext.Provider;
-import java.io.IOException;
 
 @Provider
 public class CORSFilter implements ContainerResponseFilter {
