+++ dauphine-open-data/src/main/java/io/github/oliviercailloux/y2018/opendata/provider/ResponseHeaderProvider.java
@@ -4,7 +4,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.Locale;
 
 import javax.annotation.PostConstruct;
-import javax.enterprise.context.RequestScoped;
 import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerResponseContext;
