+++ dauphine-open-data/src/main/java/io/github/oliviercailloux/y2018/opendata/provider/TraceFilterProvider.java
@@ -5,7 +5,6 @@ import java.util.Date;
 import java.util.Optional;
 
 import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerRequestFilter;
