+++ dauphine-open-data/src/test/java/io/github/oliviercailloux/y2018/opendata/provider/TraceFilterProviderIT.java
@@ -5,45 +5,26 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.net.URL;
 import java.util.Date;
 import java.util.List;
 
 import javax.inject.Inject;
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.core.Response;
 
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-import io.github.oliviercailloux.y2018.opendata.ArquillianUtils;
+import io.github.oliviercailloux.y2018.opendata.AbstractArquillianIT;
 import io.github.oliviercailloux.y2018.opendata.dao.EntityAlreadyExistsDaoException;
 import io.github.oliviercailloux.y2018.opendata.dao.HttpAuditDao;
 import io.github.oliviercailloux.y2018.opendata.entity.HttpAudit;
 import io.github.oliviercailloux.y2018.opendata.resource.Utils;
 
-@RunWith(Arquillian.class)
-public class TraceFilterProviderIT {
+public class TraceFilterProviderIT extends AbstractArquillianIT {
 
 	private static final String HEALTHCHECK_PATH = "/healthcheck";
 	
-	private Client client = ClientBuilder.newClient();
-	
 	@Inject
 	private HttpAuditDao httpAuditDao;
-	
-	@ArquillianResource
-	private URL url;
-	
-	@Deployment()
-	public static WebArchive makeWar() {
-		return ArquillianUtils.makeWar("trace-filter");
-	}
 	
 	private HttpAudit getLastInserted(List<HttpAudit> oldAudits, List<HttpAudit> newAudits) {
 		System.err.println(oldAudits);
