+++ dauphine-open-data/src/test/java/io/github/oliviercailloux/y2018/opendata/resource/CourseResourceIT.java
@@ -5,9 +5,6 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.ws.rs.core.GenericType;
 
-import org.junit.Ignore;
-import org.junit.runners.Suite.SuiteClasses;
-
 import io.github.oliviercailloux.y2018.opendata.dao.CourseDao;
 import io.github.oliviercailloux.y2018.opendata.entity.Course;
