+++ dauphine-open-data/src/test/java/io/github/oliviercailloux/y2018/opendata/resource/LectureResourceIT.java
@@ -6,8 +6,6 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.ws.rs.core.GenericType;
 
-import org.junit.Ignore;
-
 import com.diffplug.common.base.Errors;
 
 import io.github.oliviercailloux.y2018.opendata.dao.CourseDao;
